+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/relicrecoveryv2/RelicTelyMode.java
@@ -6,8 +6,17 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
  * Created by christinecarroll on 12/14/17.
  */
 
-public abstract class RelicTelyMode  extends MeccyMode{
-    abstract public void runOpMode();
+public class RelicTelyMode  extends MeccyMode{
+    public void runOpMode() {
+
+    }
+
     //
+    private double getPathagorus(double a, double b){//Define Pythagorean Theorem
+        double c = Math.sqrt(Math.pow(a, 2) + Math.pow(b, 2));
+        return c;
+
+        //Finding the joystick position using pythagorean theorem (a2 + b2 = c2)
+    }
 
 }
